add usepromotionsettings hook
